Change labor ownership behavior
In the previous way the world worked, fates were open chains
so when a fate create a labor, we decided if that labor was for
the system owner or the service owner by looking at the fates
that came afterwards.  Now, the fate that creates the labor
must decide for whom the labor exists.