Working towards new fate locked quest creation
Soon, when creating a quest, you'll be able to specify the fate
you want to start with.  This may become the actual default.

Also, update the notify script to add the owner field to the
subject line.

Finally, also added the ability to go directly to a specific
labor via URL.