Bugfix: hermes CLI labor info doesn't probably display resolutions
Currently, the hermes cli uses the prior method of determining possible closing actions where it looked for fates
that were created by the same event that created the labor.  But now we have an explicit ID of the fate that created
a labor so we don't need to do that.